+++ Source/TracerCUDA/ImageStructs.h
@@ -8,12 +8,12 @@ template<class T>
 struct ImageGMem
 {
     T*          gPixels;
-    uint32_t*   gSampleCounts;
+    float*      gSampleCounts;
 };
 
 template<class T>
 struct ImageGMemConst
 {
-    const T* gPixels;
-    const uint32_t* gSampleCounts;
+    const T*        gPixels;
+    const float*    gSampleCounts;
 };
